Organise domain logic
